Changes the default offset from "end" to "stored" when loading from
Kafka. This aligns with typical consumer expectations where reading
resumes from the last committed position rather than skipping to the end
of the topic.

Also fixes the offset parameter type from int to int64_t for consistency
with the RdKafka API.

+++ plugins/kafka/src/configuration.cpp
@@ -230,7 +230,7 @@ auto configuration::set_rebalance_cb(int64_t offset) -> caf::error {
   return {};
 }
 
-configuration::rebalancer::rebalancer(int offset) : offset_{offset} {
+configuration::rebalancer::rebalancer(int64_t offset) : offset_{offset} {
 }
 
 auto configuration::rebalancer::rebalance_cb(
